Add option to create missing subtitle request
This is needed for subtitles that were imported from YT
and marked as complete. Such subtitles are only editable by
language managers. This can be solved if we create the subtitle request
for such video for a given language.

Since this app is public, we only allow this for a selected subset of
languages.